requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
entrypoints==0.3
flake8==3.7.9
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
Janome==0.3.10
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.1
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
tqdm==4.45.0
urllib3==1.25.8
wordcloud==1.6.0
yapf==0.29.0
